backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
email-validator==2.1.0
requests==2.31.0
httpx==0.26.0
cloudinary==1.36.0

slowapi==0.1.9
bleach==6.1.0
openai==1.12.0
